// src/shared/dataprovider/api/model/merchantDTO.ts
/**
 * Dynasty Cash In Gateway
 * Dynasty Cash In API list
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ClientShortenedDTO } from './clientShortenedDTO';
import { FeeDTO } from './feeDTO';
import { MerchantVendorDTO } from './merchantVendorDTO';
import { PaymentChannelDTO } from './paymentChannelDTO';
import { StatusEnum } from './statusEnum';


export interface MerchantDTO { 
    Id?: number;
    PublicIdCashIn?: number;
    PublicIdCashOut?: number;
    Name?: string;
    TotalCashIn?: number;
    TotalCashOut?: number;
    TotalTransferredToRemittanceWallet?: number;
    Balance?: number;
    TopUpBalance?: number;
    Status?: StatusEnum;
    CreatedDate?: Date;
    LastUpdatedDate?: Date;
    CashInFee?: FeeDTO;
    CashOutFee?: FeeDTO;
    PaymentChannel?: PaymentChannelDTO;
    Vendor?: MerchantVendorDTO;
    Client?: ClientShortenedDTO;
    
    
}