// src/shared/dataprovider/api/model/cardDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CardTransactionDTO } from './cardTransactionDTO';
import { CardTypeEnum } from './cardTypeEnum';
import { ClientDTO } from './clientDTO';
import { StatusEnum } from './statusEnum';

export interface CardDTO { 
    Id?: number;
    ClientId?: number;
    Type?: CardTypeEnum;
    Balance?: number;
    Status?: StatusEnum;
    TotalWalletReleased?: number;
    CreatedDate?: Date;
    UpdatedDate?: Date;
    BankName?: string;
    AccountName?: string;
    NickName?: string;
    AccountNumber?: string;
    BinanceWalletAddress?: string;
    IsMainCard?: boolean;
    Client?: ClientDTO;
    CardTransactions?: Array<CardTransactionDTO>;
}