// src/shared/dataprovider/api/model/clientBalanceResultDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ClientBalanceDTO } from './clientBalanceDTO';

export interface ClientBalanceResultDTO { 
    TotalRecords?: number;
    TotalPages?: number;
    PageNum?: number;
    Clients?: Array<ClientBalanceDTO>;
}